refactor: make standardized for Form
