+++ packages/form/index.js
@@ -1,4 +1,4 @@
-import ElForm from './src/form'
+import ElForm from './Form.vue'
 
 /* istanbul ignore next */
 ElForm.install = function(app) {
